requirements.txt
asgiref==3.2.3
astroid==2.3.3
awsebcli==3.17.0
blessed==1.16.1
boto==2.49.0
boto3==1.10.45
botocore==1.13.45
cached-property==1.5.1
cement==2.8.2
certifi==2019.11.28
chardet==3.0.4
colorama==0.3.9
dj-database-url==0.5.0
Django==2.1.5
django-amazon-ses==2.1.1
django-ckeditor==5.4.0
django-js-asset==1.2.2
django-ses==0.8.13
django-smtp-ssl==1.0
django-storages==1.8
docker==3.7.3
docker-compose==1.23.2
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
docutils==0.15.2
drf-http-response==0.0.1.0
future==0.16.0
idna==2.7
isort==4.3.21
jmespath==0.9.4
jsonschema==2.6.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
mysql-connector-python==8.0.18
mysqlclient==1.4.6
pathspec==0.5.9
Pillow==6.2.1
protobuf==3.6.0
psycopg2-binary==2.8.4
pylint==2.4.4
pylint-django==2.0.13
pylint-plugin-utils==0.6
python-dateutil==2.8.0
pytz==2019.3
PyYAML==3.13
requests==2.20.1
s3transfer==0.2.1
semantic-version==2.5.0
six==1.14.0
sqlparse==0.3.0
termcolor==1.1.0
texttable==0.9.1
typed-ast==1.4.0
urllib3==1.24.3
wcwidth==0.1.7
websocket-client==0.57.0
wrapt==1.11.2
